Don't compile LICENSE info for installed packages in the container.
We've had a stanza that downloads a collection of source packages and includes
them directly in our containers, which dates back to the initial Dockerfile
commit in the datalab repo. Our understanding is that this is unnecessary (cf
conversation and notebook in http://b/68012493).

This just removes the steps that fetch and bundle these files, and replaces
the LICENSE file in this directory with the LICENSE for the github repo that
lives here.

RELNOTES: n/a
PiperOrigin-RevId: 183318461