Faster argo builds with better labels.
It turns out you can request to use a larger GCE VM when doing builds on
Argo; this seems to ~halve the time to build our containers (~30m to ~15m),
which is pure win.

I also added an `id` to each of our build steps; this makes the output
clearer, and is required if/when we sequence multiple build steps.

RELNOTES: n/a
PiperOrigin-RevId: 195371319