Setting LANG env variable before attempting to run tox unit tests.
Why? Colabtools tests require a UTF-8 environment. We've done the work to make this the case when building a Docker container, but the environment variable was being set after tests had run. This simply moves the environment variable declaration to occur before tests are run.

Tested: Ran update.sh and validated that build step completed successfully. Also patched a rollback of cl/205448253 and ensured that tests succeeded.

RELNOTES: n/a
PiperOrigin-RevId: 206630628